+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/adapter/view/GiphyMediaAttachmentViewStyle.kt
@@ -5,9 +5,7 @@ import android.graphics.drawable.Drawable
 import android.util.AttributeSet
 import android.widget.ImageView
 import androidx.annotation.ColorInt
-import androidx.annotation.DimenRes
 import io.getstream.chat.android.ui.R
-import io.getstream.chat.android.ui.common.extensions.internal.dpToPx
 import io.getstream.chat.android.ui.common.extensions.internal.getColorCompat
 import io.getstream.chat.android.ui.common.extensions.internal.getDrawableCompat
 import io.getstream.chat.android.ui.common.extensions.internal.getEnum
@@ -22,7 +20,6 @@ import io.getstream.chat.android.ui.message.list.adapter.view.internal.GiphyMedi
  * @param giphyIcon Displays the Giphy logo over the Giphy image.
  * @param placeholderIcon Displayed while the Giphy is Loading.
  * @param imageBackgroundColor Sets the background colour for the Giphy container.
- * @param giphyMaxHeight Sets the maximum height a Giphy container is allowed to have.
  * @param giphyType Sets the Giphy type which directly affects image quality and if the container is resized or not.
  * @param scaleType Sets the scaling type for loading the image. E.g. 'centerCrop', 'fitCenter', etc...
  */
@@ -31,13 +28,10 @@ public class GiphyMediaAttachmentViewStyle(
     public val giphyIcon: Drawable,
     public val placeholderIcon: Drawable,
     @ColorInt public val imageBackgroundColor: Int,
-    @DimenRes public val giphyMaxHeight: Int,
     public val giphyType: GiphyMediaAttachmentView.Companion.GiphyInfoType,
     public val scaleType: ImageView.ScaleType,
 ) {
     internal companion object {
-        private const val DEFAULT_HEIGHT_DP = 200
-
         /**
          * Fetches styled attributes and returns them wrapped inside of [GiphyMediaAttachmentViewStyle].
          */
@@ -65,35 +59,23 @@ public class GiphyMediaAttachmentViewStyle(
                     attributes.getDrawable(R.styleable.GiphyMediaAttachmentView_streamUiGiphyMediaAttachmentPlaceHolderIcon)
                         ?: context.getDrawableCompat(R.drawable.stream_ui_picture_placeholder)!!
 
-                val giphyHeight =
-                    attributes.getDimensionPixelSize(R.styleable.GiphyMediaAttachmentView_streamUiGiphyMediaAttachmentGiphyMaxHeight,
-                        DEFAULT_HEIGHT_DP.dpToPx())
-
                 val giphyType =
                     attributes.getEnum(R.styleable.GiphyMediaAttachmentView_streamUiGiphyMediaAttachmentGiphyType,
                         GiphyMediaAttachmentView.Companion.GiphyInfoType.FIXED_HEIGHT)
 
                 val scaleType =
                     attributes.getEnum(R.styleable.GiphyMediaAttachmentView_streamUiGiphyMediaAttachmentScaleType,
-                        ImageView.ScaleType.CENTER_CROP)
+                        ImageView.ScaleType.FIT_CENTER)
 
                 return GiphyMediaAttachmentViewStyle(
                     progressIcon = progressIcon,
                     giphyIcon = giphyIcon,
                     placeholderIcon = placeholderIcon,
                     imageBackgroundColor = imageBackgroundColor,
-                    giphyMaxHeight = giphyHeight,
                     giphyType = giphyType,
                     scaleType = scaleType
                 )
             }
         }
     }
-
-    /**
-     * Returns if the giphy should use a fixed constant size (height) or if it should be loaded and resized based on the
-     * GIF.
-     */
-    public fun isConstantSizeEnabled(): Boolean =
-        this.giphyType == GiphyMediaAttachmentView.Companion.GiphyInfoType.FIXED_HEIGHT || this.giphyType == GiphyMediaAttachmentView.Companion.GiphyInfoType.FIXED_HEIGHT_DOWNSAMPLED
 }